# openapi.yml
openapi: 3.0.0
info:
  title: My API
  version: 1.0.0

servers:
  - url: http://localhost:8090
    description: commande Service

paths:
  /commandes:
    get:
      tags:
        - commandes
      summary: Get a list of commands
      description: Retrieve information about commandes.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                    desc:
                      type: string
                    quantity:
                      type: integer
                      format: int32
                    date:
                      type: string
                      format: date
                    amount:
                      type: number
                      format: double

    post:
      tags:
        - commandes
      summary: Create a new command
      description: Add a new command to the list.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                desc:
                  type: string
                quantity:
                  type: integer
                  format: int32
                date:
                  type: string
                  format: date
                amount:
                  type: number
                  format: double
      responses:
        '201':
          description: Command created successfully
        '400':
          description: Bad request, invalid input

  /commandes/{id}:
    get:
      tags:
        - commandes
      summary: Get command by ID
      description: Retrieve information about a specific command.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the command
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  desc:
                    type: string
                  quantity:
                    type: integer
                    format: int32
                  date:
                    type: string
                    format: date
                  amount:
                    type: number
                    format: double
        '404':
          description: Command not found

    put:
      tags:
        - commandes
      summary: Update command by ID
      description: Update information about a specific command.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the command
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                desc:
                  type: string
                quantity:
                  type: integer
                  format: int32
                date:
                  type: string
                  format: date
                amount:
                  type: number
                  format: double
      responses:
        '200':
          description: Successful response
        '400':
          description: Bad request, invalid input
        '404':
          description: Command not found

    delete:
      tags:
        - commandes
      summary: Delete command by ID
      description: Remove a specific command.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the command
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Command deleted successfully
        '404':
          description: Command not found